integrate react router into app and use google fonts
